+++ libcloud/test/compute/test_deployment.py
@@ -24,6 +24,7 @@ import unittest
 from libcloud.utils.py3 import httplib
 from libcloud.utils.py3 import u
 from libcloud.utils.py3 import PY3
+from libcloud.utils.py3 import assertRaisesRegex
 
 from libcloud.compute.deployment import MultiStepDeployment, Deployment
 from libcloud.compute.deployment import SSHKeyDeployment, ScriptDeployment
@@ -32,6 +33,8 @@ from libcloud.compute.base import Node
 from libcloud.compute.base import NodeAuthPassword
 from libcloud.compute.types import NodeState, DeploymentError, LibcloudError
 from libcloud.compute.ssh import BaseSSHClient
+from libcloud.compute.ssh import have_paramiko
+from libcloud.compute.ssh import SSHCommandTimeoutError
 from libcloud.compute.drivers.rackspace import RackspaceFirstGenNodeDriver as Rackspace
 
 from libcloud.test import MockHttp, XML_HEADERS
@@ -57,15 +60,19 @@ class MockDeployment(Deployment):
 
 class MockClient(BaseSSHClient):
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, throw_on_timeout=False, *args, **kwargs):
         self.stdout = ''
         self.stderr = ''
         self.exit_status = 0
+        self.throw_on_timeout = throw_on_timeout
 
     def put(self, path, contents, chmod=755, mode='w'):
         return contents
 
-    def run(self, name):
+    def run(self, cmd, timeout=None):
+        if self.throw_on_timeout and timeout is not None:
+            raise ValueError("timeout")
+
         return self.stdout, self.stderr, self.exit_status
 
     def delete(self, name):
@@ -117,14 +124,25 @@ class DeploymentTests(unittest.TestCase):
         sd2 = ScriptDeployment(script='foobar', delete=False)
         sd3 = ScriptDeployment(
             script='foobar', delete=False, name='foobarname')
+        sd4 = ScriptDeployment(
+            script='foobar', delete=False, name='foobarname', timeout=10)
 
         self.assertTrue(sd1.name.find('deployment') != '1')
         self.assertEqual(sd3.name, 'foobarname')
+        self.assertEqual(sd3.timeout, None)
+        self.assertEqual(sd4.timeout, 10)
 
         self.assertEqual(self.node, sd1.run(node=self.node,
                                             client=MockClient(hostname='localhost')))
         self.assertEqual(self.node, sd2.run(node=self.node,
                                             client=MockClient(hostname='localhost')))
+        self.assertEqual(self.node, sd3.run(node=self.node,
+                                            client=MockClient(hostname='localhost')))
+
+        assertRaisesRegex(self, ValueError, 'timeout', sd4.run,
+                          node=self.node,
+                          client=MockClient(hostname='localhost',
+                                            throw_on_timeout=True))
 
     def test_script_file_deployment(self):
         file_path = os.path.abspath(__file__)
@@ -136,6 +154,10 @@ class DeploymentTests(unittest.TestCase):
 
         sfd1 = ScriptFileDeployment(script_file=file_path)
         self.assertEqual(sfd1.script, content)
+        self.assertEqual(sfd1.timeout, None)
+
+        sfd2 = ScriptFileDeployment(script_file=file_path, timeout=20)
+        self.assertEqual(sfd2.timeout, 20)
 
     def test_script_deployment_relative_path(self):
         client = Mock()
@@ -145,7 +167,7 @@ class DeploymentTests(unittest.TestCase):
         sd = ScriptDeployment(script='echo "foo"', name='relative.sh')
         sd.run(self.node, client)
 
-        client.run.assert_called_once_with(FILE_PATH)
+        client.run.assert_called_once_with(FILE_PATH, timeout=None)
 
     def test_script_deployment_absolute_path(self):
         client = Mock()
@@ -157,7 +179,7 @@ class DeploymentTests(unittest.TestCase):
         sd = ScriptDeployment(script='echo "foo"', name=file_path)
         sd.run(self.node, client)
 
-        client.run.assert_called_once_with(file_path)
+        client.run.assert_called_once_with(file_path, timeout=None)
 
     def test_script_deployment_with_arguments(self):
         client = Mock()
@@ -172,7 +194,7 @@ class DeploymentTests(unittest.TestCase):
         sd.run(self.node, client)
 
         expected = '%s arg1 arg2 --option1=test' % (file_path)
-        client.run.assert_called_once_with(expected)
+        client.run.assert_called_once_with(expected, timeout=None)
 
         client.reset_mock()
 
@@ -182,7 +204,7 @@ class DeploymentTests(unittest.TestCase):
         sd.run(self.node, client)
 
         expected = file_path
-        client.run.assert_called_once_with(expected)
+        client.run.assert_called_once_with(expected, timeout=None)
 
     def test_script_file_deployment_with_arguments(self):
         file_path = os.path.abspath(__file__)
@@ -197,7 +219,7 @@ class DeploymentTests(unittest.TestCase):
         sfd.run(self.node, client)
 
         expected = '%s arg1 arg2 --option1=test option2' % (file_path)
-        client.run.assert_called_once_with(expected)
+        client.run.assert_called_once_with(expected, timeout=None)
 
     def test_script_deployment_and_sshkey_deployment_argument_types(self):
         class FileObject(object):
@@ -362,6 +384,31 @@ class DeploymentTests(unittest.TestCase):
         else:
             self.fail('Exception was not thrown')
 
+    @unittest.skipIf(not have_paramiko, 'Skipping because paramiko is not available')
+    def test_ssh_client_connect_immediately_throws_on_fatal_execption(self):
+        # Verify that fatal exceptions are immediately propagated and ensure
+        # we don't try to retry on them
+        from paramiko.ssh_exception import SSHException
+        from paramiko.ssh_exception import PasswordRequiredException
+
+        mock_ssh_client = Mock()
+        mock_ssh_client.connect = Mock()
+        mock_ssh_client.connect.side_effect = IOError('bam')
+
+        mock_exceptions = [
+            SSHException('Invalid or unsupported key type'),
+            PasswordRequiredException('private key file is encrypted'),
+            SSHException('OpenSSH private key file checkints do not match')
+        ]
+
+        for mock_exception in mock_exceptions:
+            mock_ssh_client.connect = Mock(side_effect=mock_exception)
+            assertRaisesRegex(self, mock_exception.__class__, str(mock_exception),
+                              self.driver._ssh_client_connect,
+                              ssh_client=mock_ssh_client,
+                              wait_period=0.1,
+                              timeout=0.2)
+
     def test_run_deployment_script_success(self):
         task = Mock()
         ssh_client = Mock()
@@ -388,6 +435,23 @@ class DeploymentTests(unittest.TestCase):
         else:
             self.fail('Exception was not thrown')
 
+    def test_run_deployment_script_ssh_command_timeout_fatal_exception(self):
+        # We shouldn't retry on SSHCommandTimeoutError error since it's fatal
+        task = Mock()
+        task.run = Mock()
+        task.run.side_effect = SSHCommandTimeoutError('ls -la', 10)
+        ssh_client = Mock()
+
+        try:
+            self.driver._run_deployment_script(task=task,
+                                               node=self.node,
+                                               ssh_client=ssh_client,
+                                               max_tries=5)
+        except SSHCommandTimeoutError as e:
+            self.assertTrue(e.message.find('Command didn\'t finish') != -1)
+        else:
+            self.fail('Exception was not thrown')
+
     @patch('libcloud.compute.base.SSHClient')
     @patch('libcloud.compute.ssh')
     def test_deploy_node_success(self, mock_ssh_module, _):
@@ -399,6 +463,64 @@ class DeploymentTests(unittest.TestCase):
 
         node = self.driver.deploy_node(deploy=deploy)
         self.assertEqual(self.node.id, node.id)
+
+    @patch('libcloud.compute.base.atexit')
+    @patch('libcloud.compute.base.SSHClient')
+    @patch('libcloud.compute.ssh')
+    def test_deploy_node_at_exit_func_functionality(self, mock_ssh_module, _, mock_at_exit):
+        self.driver.create_node = Mock()
+        self.driver.create_node.return_value = self.node
+        mock_ssh_module.have_paramiko = True
+
+        deploy = Mock()
+
+        def mock_at_exit_func(driver, node):
+            pass
+
+        # On success, at exit handler should be unregistered
+        self.assertEqual(mock_at_exit.register.call_count, 0)
+        self.assertEqual(mock_at_exit.unregister.call_count, 0)
+
+        node = self.driver.deploy_node(deploy=deploy, at_exit_func=mock_at_exit_func)
+        self.assertEqual(mock_at_exit.register.call_count, 1)
+        self.assertEqual(mock_at_exit.unregister.call_count, 1)
+        self.assertEqual(self.node.id, node.id)
+
+        # On deploy failure, at exit handler should also be unregistered
+        mock_at_exit.reset_mock()
+
+        deploy.run.side_effect = Exception('foo')
+
+        self.assertEqual(mock_at_exit.register.call_count, 0)
+        self.assertEqual(mock_at_exit.unregister.call_count, 0)
+
+        try:
+            self.driver.deploy_node(deploy=deploy, at_exit_func=mock_at_exit_func)
+        except DeploymentError as e:
+            self.assertTrue(e.node.id, self.node.id)
+        else:
+            self.fail('Exception was not thrown')
+
+        self.assertEqual(mock_at_exit.register.call_count, 1)
+        self.assertEqual(mock_at_exit.unregister.call_count, 1)
+
+        # But it should not be registered on create_node exception
+        mock_at_exit.reset_mock()
+
+        self.driver.create_node = Mock(side_effect=Exception('Failure'))
+
+        self.assertEqual(mock_at_exit.register.call_count, 0)
+        self.assertEqual(mock_at_exit.unregister.call_count, 0)
+
+        try:
+            self.driver.deploy_node(deploy=deploy, at_exit_func=mock_at_exit_func)
+        except Exception as e:
+            self.assertTrue('Failure' in str(e))
+        else:
+            self.fail('Exception was not thrown')
+
+        self.assertEqual(mock_at_exit.register.call_count, 0)
+        self.assertEqual(mock_at_exit.unregister.call_count, 0)
 
     @patch('libcloud.compute.base.SSHClient')
     @patch('libcloud.compute.ssh')
@@ -440,6 +562,7 @@ class DeploymentTests(unittest.TestCase):
         # the arguments
         global call_count
         call_count = 0
+
         def create_node(name, image, size, ex_custom_arg_1, ex_custom_arg_2,
                         ex_foo=None, auth=None, **kwargs):
             global call_count